Wanky connection to the rendering stuff. However, following problems. - Because rasterize_point.h/cu are not in the CMakeLists.txt included in the library build, I had on my side linking problem. For now I just have copied those files into my src and include directory. - The return values of the rendering might dont match what I actually need. I will need a solution here

+++ includes/camera.cuh
@@ -143,8 +143,8 @@ public:
     uint32_t _camera_ID;
     Eigen::Matrix3d _R;
     Eigen::Vector3d _T;
-    double _fov_x;
-    double _fov_y;
+    float _fov_x;
+    float _fov_y;
     std::string _image_name;
     std::string _image_path;
     CAMERA_MODEL _camera_model;
